+++ Minifiers/jsonMinifier.js
@@ -1,26 +1,79 @@
 /**
- * @file That is called when the user minifies a css document
- * This file receives the css document and returns the minified one.
+ * @file That is called when the user minifies a json document
+ * This file receives the json document and returns the minified one.
  * @author Jose Gracia Berenguer
- * @since 0.1.0
+ * @since 0.2.0
  * @see ../README.md
  * @link https://github.com/Josee9988/MinifyAll
  */
+
 class Minifier {
     /**
      * Minifier constructor that maps and trims the code.
-     * @param {Array} cssContent all the code that will be minified 
+     * @param {Array} jsonContent all the code that will be minified 
      */
-    constructor(cssContent) {
-        this.cssContent = cssContent.map(content => content.trim());
+    constructor(jsonContent) {
+        this.jsonContent = jsonContent.map(content => content.trim());
     }
 
     /**
-     * getCssMinified finds lasts spaces and trim it into just one line
+     * getJSONMinified finds lasts spaces and trim it into just one line
      * @return {String} the line compressed
      */
-    getCssMinified() {
-        return this.cssContent.join('').replace(/;\}|\s+}/g, '}').replace(/\/\*.*?\*\//g, '').replace(/:\s/g, ':').replace(/ {/g, '{').replace(/[ \t]/g, '');
+    getJSONMinified() {
+        return this.jsonContent.join('').replace(/;\}|\s+}/g, '}').replace(/\/\*.*?\*\//g, '').replace(/:\s/g, ':').replace(/ {/g, '{').replace(/[ \t]/g, '');
+    }
+
+    /**
+     * removeSingleLineComments checks line by line 
+     * and removes the single lines comments //
+     */
+    removeSingleLineComments() {
+        for (let i = 0; i < this.jsonContent.length; i++) {
+            let hexadecimal = this.jsonContent[i].replace(/[/]+.*/g, '');
+            if (hexadecimal != null) {
+                this.jsonContent[i].replace(/[/]+.*/g, '');
+                this.jsonContent[i] = hexadecimal;
+            }
+        }
+    }
+
+    /**
+     * removeMultipleLineComments checks line by line
+     * and removes the multiple line comments
+     * it only removes the content inside the comments
+     * if the multiple line comment is placed in a line
+     * with usefull code it will not be replaced.
+     */
+    removeMultipleLineComments() {
+        console.log("Multi line comments;");
+        for (let i = 0; i < this.jsonContent.length; i++) {
+            let begin = this.jsonContent[i].match(/(\/\*)/ig); // first /* found
+            if (begin != null) {
+                for (let j = 0; j < this.jsonContent.length; j++) {
+                    let end = this.jsonContent[j].match(/(\*\/)/g); //found */ end
+                    if (end != null) {
+                        for (let k = i; k < j + 1; k++) {
+                            if (k == i) {
+                                let FirstCharacterToRemove = this.jsonContent[k].indexOf("/*");
+                                let firstLineToReplace = this.jsonContent[k].substring(0, FirstCharacterToRemove);
+                                this.jsonContent[k] = firstLineToReplace;
+                            } else if (k == j) {
+                                let lastCharacterToRemove = this.jsonContent[k].indexOf("*/");
+                                let lasttLineToReplace = this.jsonContent[k].substring(lastCharacterToRemove + 2, this.jsonContent[k].length);
+                                this.jsonContent[k] = lasttLineToReplace;
+
+                            } else {
+                                this.jsonContent[k] = '';
+
+                            }
+                        }
+                        break;
+                    }
+
+                }
+            }
+        }
     }
 
     /**
@@ -30,13 +83,13 @@ class Minifier {
      * original line with the shortened one.
      */
     shortHexMain() {
-        for (let i = 0; i < this.cssContent.length; i++) {
-            let hexadecimal = this.cssContent[i].match(/#[a-f\d]{6}/ig);
+        for (let i = 0; i < this.jsonContent.length; i++) {
+            let hexadecimal = this.jsonContent[i].match(/#[a-f\d]{6}/ig);
             if (hexadecimal != null) {
                 let hexadecimalString = hexadecimal.toString();
                 let shortHex = this.getShortHexColorCode(hexadecimalString);
-                let newShortString = this.cssContent[i].replace(hexadecimalString, shortHex);
-                this.cssContent[i] = newShortString;
+                let newShortString = this.jsonContent[i].replace(hexadecimalString, shortHex);
+                this.jsonContent[i] = newShortString;
             }
         }
     }
@@ -48,15 +101,14 @@ class Minifier {
      * Then it replaces the original line with the shortened one.
      */
     shortRGBMain() {
-        for (let i = 0; i < this.cssContent.length; i++) {
-            let rgb = this.cssContent[i].match(/((rgb)\((\d{1,3}%?,\s?){3}(1|0?\.\d+)\)|(rgb)\(\d{1,3}%?(,\s?\d{1,3}%?){2}\))/ig);
+        for (let i = 0; i < this.jsonContent.length; i++) {
+            let rgb = this.jsonContent[i].match(/((rgb)\((\d{1,3}%?,\s?){3}(1|0?\.\d+)\)|(rgb)\(\d{1,3}%?(,\s?\d{1,3}%?){2}\))/ig);
             if (rgb != null) {
                 let rgbString = rgb.toString();
                 let result = this.rgbArrayToObject(rgbString);
                 let shortHex = this.rgbToShortHex(result);
-                let newShortString = this.cssContent[i].replace(rgbString, shortHex);
-                this.cssContent[i] = newShortString;
-
+                let newShortString = this.jsonContent[i].replace(rgbString, shortHex);
+                this.jsonContent[i] = newShortString;
             }
         }
     }
@@ -120,7 +172,5 @@ class Minifier {
         var hexB = Math.round(rgb.b / 17).toString(16);
         return "#" + hexR + "" + hexG + "" + hexB;
     }
-
 }
-
 module.exports = Minifier;